// app/music-app/src/app/enums/srv-paths.enum.ts
/* eslint-disable @typescript-eslint/naming-convention */
export enum API_PATHS {
  FAVORITES = 'app/Favorites',
  PERSONALIZATIONS = 'app/Personalizations',
  PERSONALIZATIONS_FAVORITES = 'favorites',
  PERSONALIZATIONS_CUSTOM_PLAYLISTS = 'customPlaylists',
  CUSTOM_PLAYLISTS = 'app/CustomPlaylists',
  CUSTOM_PLAYLISTS_TRACKS = 'tracks',
}